feat: restructure sizzle app by organizing views, urls, and admin registrations
