+++ blt/urls.py
@@ -213,18 +213,13 @@ from website.views.organization import (
     RoomCreateView,
     RoomsListView,
     ScoreboardView,
-    TimeLogListAPIView,
-    TimeLogListView,
     UpcomingHunts,
     add_domain_to_organization,
     add_or_update_domain,
     add_or_update_organization,
-    add_sizzle_checkIN,
     admin_organization_dashboard,
     admin_organization_dashboard_detail,
     approve_activity,
-    checkIN,
-    checkIN_detail,
     delete_room,
     delete_time_entry,
     dislike_activity,
@@ -241,15 +236,11 @@ from website.views.organization import (
     organization_hunt_results,
     room_messages_api,
     send_message_api,
-    sizzle,
-    sizzle_daily_log,
-    sizzle_docs,
     subscribe_to_domains,
     trademark_detailview,
     trademark_search,
     update_organization_repos,
     update_role,
-    user_sizzle_report,
     view_hunt,
     weekly_report,
 )
@@ -968,19 +959,7 @@ urlpatterns = [
         TagApiViewset.as_view({"get": "list", "post": "create"}),
         name="tags-api",
     ),
-    path("sizzle/", sizzle, name="sizzle"),
-    path("check-in/", checkIN, name="checkIN"),
-    path("add-sizzle-checkin/", add_sizzle_checkIN, name="add_sizzle_checkin"),
-    path("check-in/<int:report_id>/", checkIN_detail, name="checkIN_detail"),
-    path("sizzle-docs/", sizzle_docs, name="sizzle-docs"),
-    path("api/timelogsreport/", TimeLogListAPIView, name="timelogsreport"),
-    path("time-logs/", TimeLogListView, name="time_logs"),
-    path("sizzle-daily-log/", sizzle_daily_log, name="sizzle_daily_log"),
-    path(
-        "user-sizzle-report/<str:username>/",
-        user_sizzle_report,
-        name="user_sizzle_report",
-    ),
+    path("sizzle/", include("sizzle.urls")),
     path("submit-roadmap-pr/", submit_roadmap_pr, name="submit-roadmap-pr"),
     path("view-pr-analysis/", view_pr_analysis, name="view_pr_analysis"),
     path("delete_time_entry/", delete_time_entry, name="delete_time_entry"),
