+++ website/views/organization.py
@@ -3,7 +3,6 @@ import json
 import logging
 import re
 import time
-from collections import defaultdict
 from datetime import datetime, timedelta
 from decimal import Decimal
 from urllib.parse import urlparse
@@ -19,26 +18,16 @@ from django.core.cache import cache
 from django.core.mail import send_mail
 from django.core.paginator import EmptyPage, PageNotAnInteger, Paginator
 from django.db.models import Count, Prefetch, Q, Sum
-from django.http import (
-    Http404,
-    HttpResponse,
-    HttpResponseBadRequest,
-    HttpResponseRedirect,
-    JsonResponse,
-    StreamingHttpResponse,
-)
+from django.http import Http404, HttpResponse, HttpResponseRedirect, JsonResponse, StreamingHttpResponse
 from django.shortcuts import get_object_or_404, redirect, render
 from django.urls import reverse, reverse_lazy
 from django.utils import timezone
-from django.utils.dateparse import parse_datetime
 from django.utils.decorators import method_decorator
 from django.utils.text import slugify
 from django.utils.timezone import now
 from django.views.decorators.http import require_POST
 from django.views.generic import DetailView, FormView, ListView, TemplateView, View
 from django.views.generic.edit import CreateView
-from rest_framework import status
-from rest_framework.authtoken.models import Token
 
 from website.forms import CaptchaForm, HuntForm, IpReportForm, RoomForm, UserProfileForm
 from website.models import (
@@ -65,7 +54,7 @@ from website.models import (
     Winner,
 )
 from website.services.blue_sky_service import BlueSkyService
-from website.utils import format_timedelta, get_client_ip, get_github_issue_title, rebuild_safe_url, validate_file_type
+from website.utils import get_client_ip, rebuild_safe_url, validate_file_type
 
 logger = logging.getLogger(__name__)
 
@@ -1116,223 +1105,6 @@ class CreateHunt(TemplateView):
                 return HttpResponse("failed")
         except (OrganizationAdmin.DoesNotExist, Domain.DoesNotExist, ValueError, KeyError) as e:
             return HttpResponse(f"Error: {str(e)}")
-
-
-@login_required
-def user_sizzle_report(request, username):
-    user = get_object_or_404(User, username=username)
-    time_logs = TimeLog.objects.filter(user=user).order_by("-start_time")
-
-    grouped_logs = defaultdict(list)
-    for log in time_logs:
-        date_str = log.created.strftime("%Y-%m-%d")
-        grouped_logs[date_str].append(log)
-
-    response_data = []
-    for date, logs in grouped_logs.items():
-        first_log = logs[0]
-        total_duration = sum((log.duration for log in logs if log.duration), timedelta())
-
-        total_duration_seconds = total_duration.total_seconds()
-        formatted_duration = f"{int(total_duration_seconds // 60)} min {int(total_duration_seconds % 60)} sec"
-
-        issue_title = get_github_issue_title(first_log.github_issue_url)
-
-        start_time = first_log.start_time.strftime("%I:%M %p")
-        end_time = first_log.end_time.strftime("%I:%M %p")
-        formatted_date = first_log.created.strftime("%d %B %Y")
-
-        day_data = {
-            "issue_title": issue_title,
-            "duration": formatted_duration,
-            "start_time": start_time,
-            "end_time": end_time,
-            "date": formatted_date,
-        }
-
-        response_data.append(day_data)
-
-    return render(
-        request,
-        "sizzle/user_sizzle_report.html",
-        {"response_data": response_data, "user": user},
-    )
-
-
-@login_required
-def sizzle_daily_log(request):
-    try:
-        if request.method == "GET":
-            reports = DailyStatusReport.objects.filter(user=request.user).order_by("-date")
-            return render(request, "sizzle/sizzle_daily_status.html", {"reports": reports})
-
-        if request.method == "POST":
-            previous_work = request.POST.get("previous_work")
-            next_plan = request.POST.get("next_plan")
-            blockers = request.POST.get("blockers")
-            goal_accomplished = request.POST.get("goal_accomplished") == "on"
-            current_mood = request.POST.get("feeling")
-            print(previous_work, next_plan, blockers, goal_accomplished, current_mood)
-
-            DailyStatusReport.objects.create(
-                user=request.user,
-                date=now().date(),
-                previous_work=previous_work,
-                next_plan=next_plan,
-                blockers=blockers,
-                goal_accomplished=goal_accomplished,
-                current_mood=current_mood,
-            )
-
-            messages.success(request, "Daily status report submitted successfully.")
-            return JsonResponse(
-                {
-                    "success": "true",
-                    "message": "Daily status report submitted successfully.",
-                }
-            )
-
-    except Exception as e:
-        messages.error(request, f"An error occurred: {e}")
-        return redirect("sizzle")
-
-    return HttpResponseBadRequest("Invalid request method.")
-
-
-@login_required
-def TimeLogListView(request):
-    time_logs = TimeLog.objects.filter(user=request.user).order_by("-start_time")
-    active_time_log = time_logs.filter(end_time__isnull=True).first()
-
-    # print the all details of the active time log
-    token, created = Token.objects.get_or_create(user=request.user)
-    organizations_list_queryset = Organization.objects.all().values("url", "name")
-    organizations_list = list(organizations_list_queryset)
-    organization_url = None
-    if active_time_log and active_time_log.organization:
-        organization_url = active_time_log.organization.url
-    return render(
-        request,
-        "sizzle/time_logs.html",
-        {
-            "time_logs": time_logs,
-            "active_time_log": active_time_log,
-            "token": token.key,
-            "organizations_list": organizations_list,
-            "organization_url": organization_url,
-        },
-    )
-
-
-def TimeLogListAPIView(request):
-    if not request.user.is_authenticated:
-        return JsonResponse({"error": "Unauthorized"}, status=status.HTTP_401_UNAUTHORIZED)
-
-    start_date_str = request.GET.get("start_date")
-    end_date_str = request.GET.get("end_date")
-
-    if not start_date_str or not end_date_str:
-        return JsonResponse(
-            {"error": "Both start_date and end_date are required."},
-            status=status.HTTP_400_BAD_REQUEST,
-        )
-
-    start_date = parse_datetime(start_date_str)
-    end_date = parse_datetime(end_date_str)
-
-    if not start_date or not end_date:
-        return JsonResponse({"error": "Invalid date format."}, status=status.HTTP_400_BAD_REQUEST)
-
-    time_logs = TimeLog.objects.filter(user=request.user, created__range=[start_date, end_date]).order_by("created")
-
-    grouped_logs = defaultdict(list)
-    for log in time_logs:
-        date_str = log.created.strftime("%Y-%m-%d")
-        grouped_logs[date_str].append(log)
-
-    response_data = []
-    for date, logs in grouped_logs.items():
-        first_log = logs[0]
-        total_duration = sum((log.duration for log in logs if log.duration), timedelta())
-
-        total_duration_seconds = total_duration.total_seconds()
-        formatted_duration = f"{int(total_duration_seconds // 60)} min {int(total_duration_seconds % 60)} sec"
-
-        issue_title = get_github_issue_title(first_log.github_issue_url)
-
-        start_time = first_log.start_time.strftime("%I:%M %p")
-        formatted_date = first_log.created.strftime("%d %B %Y")
-
-        day_data = {
-            "id": first_log.id,
-            "issue_title": issue_title,
-            "duration": formatted_duration,
-            "start_time": start_time,
-            "date": formatted_date,
-        }
-
-        response_data.append(day_data)
-
-    return JsonResponse(response_data, safe=False, status=status.HTTP_200_OK)
-
-
-def sizzle_docs(request):
-    return render(request, "sizzle/sizzle_docs.html")
-
-
-def sizzle(request):
-    # Aggregate leaderboard data: username and total_duration
-    leaderboard_qs = (
-        TimeLog.objects.values("user__username").annotate(total_duration=Sum("duration")).order_by("-total_duration")
-    )
-
-    # Process leaderboard to include formatted_duration
-    leaderboard = []
-    for entry in leaderboard_qs:
-        username = entry["user__username"]
-        total_duration = entry["total_duration"] or timedelta()  # Handle None
-        formatted_duration = format_timedelta(total_duration)
-        leaderboard.append(
-            {
-                "username": username,
-                "formatted_duration": formatted_duration,
-            }
-        )
-
-    # Initialize sizzle_data
-    sizzle_data = None
-
-    if request.user.is_authenticated:
-        last_data = TimeLog.objects.filter(user=request.user).order_by("-created").first()
-
-        if last_data:
-            all_data = TimeLog.objects.filter(user=request.user, created__date=last_data.created.date()).order_by(
-                "created"
-            )
-
-            total_duration = sum((entry.duration for entry in all_data if entry.duration), timedelta())
-
-            formatted_duration = format_timedelta(total_duration)
-
-            github_issue_url = all_data.first().github_issue_url
-            issue_title = get_github_issue_title(github_issue_url)
-
-            start_time = all_data.first().start_time.strftime("%I:%M %p")
-            date = last_data.created.strftime("%d %B %Y")
-
-            sizzle_data = {
-                "id": last_data.id,
-                "issue_title": issue_title,
-                "duration": formatted_duration,
-                "start_time": start_time,
-                "date": date,
-            }
-
-    return render(
-        request,
-        "sizzle/sizzle.html",
-        {"sizzle_data": sizzle_data, "leaderboard": leaderboard},
-    )
 
 
 def trademark_detailview(request, slug):
